Perform sign context operations from an interface.
The current SigningContext is geared very much toward performing signing
and validation operations with a certificate whose private key is known.
This changes the requirement to be anything that implements the
interface ISigningContext. This will allow other signing providers, such
as Azure Key Vault. The longer term goal is to provide a factory
mechanism that will provide the right signing context by what is
performing the signing.

Prepwork for #31.
