+++ .editorconfig
@@ -1,0 +1,24 @@
+root = true
+
+[Makefile]
+indent_style = tab
+
+[*.{yml,json}]
+end_of_line = lf
+indent_style = space
+indent_size = 2
+
+[*.{cs,md,sh}]
+end_of_line = lf
+indent_style = space
+indent_size = 4
+
+[*.cs]
+trim_trailing_whitespace = true
+end_of_line = lf
+
+[*.csproj]
+end_of_line = lf
+
+[*.sln]
+end_of_line = lf
