+++ src/OpenVsixSignTool.Core/ECDsaCurve.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenVsixSignTool.Core
+{
+    public enum ECDsaCurve
+    {
+        p256,
+        p384,
+        p521
+    }
+}
