+++ src/OpenVsixSignTool.Core/SigningAlgorithm.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenVsixSignTool.Core
+{
+    public enum SigningAlgorithm
+    {
+        Unkonwn = 0,
+        RSA,
+        ECDSA
+    }
+}
